requirements.txt
streamlit
pandas
plotlystreamlit run app.py